pyproject.toml
[project]
name = "ner-finetuning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.10.1",
    "bitsandbytes>=0.48.1",
    "datasets>=4.2.0",
    "fastapi>=0.120.0",
    "fuzzywuzzy>=0.18.0",
    "gguf>=0.17.1",
    "huggingface-hub>=0.35.3",
    "llama-cpp-python>=0.3.16",
    "nbformat>=5.10.4",
    "pandas>=2.3.3",
    "peft>=0.17.1",
    "pillow>=12.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "transformers>=4.57.1",
    "trl>=0.12.0",
    "unsloth>=2025.7.2",
    "uvicorn[standard]>=0.38.0",
    "wandb>=0.22.2",
]
